WhatsApp Plus mod APK is now free to download
WhatsApp Plus is a third-party modification of the popular messaging app WhatsApp. It offers additional features and customizations that are not available in the official version of WhatsApp. It is not clear who developed WhatsApp Plus or when it was first released. WhatsApp Plus was likely developed by a group or individual who wanted to offer additional features and customizations for WhatsApp users.  Some of the features that have been reported to be included in WhatsApp Plus include the ability to change the appearance of the app, customize notifications, and use additional emoticons. This app is available  for free at the website Forelinks Tech:
https://forelink.blogspot.com/2022/12/waplus.html